requirements.txt
httpx
anthropic
resend
python-dotenv
beautifulsoup4
